debugged brickwallLimiter

+++ BrickwallLimiter.h
@@ -7,14 +7,14 @@ class BrickwallLimiter
 {
 public:
     BrickwallLimiter();
-    BrickwallLimiter(float sampleRate);
+    BrickwallLimiter(double sampleRate);
 
-    void prepareToPlay(float sampleRate, int nrofchannels);
-    int processSamples(std::vector<std::vector<float>>& data);
+    void prepareToPlay(double sampleRate, int nrofchannels);
+    int processSamples(std::vector<std::vector<double>>& data);
 
-    void setGainLimit(float newLimit){m_Limit = newLimit;};
-    void setReleaseTime(float releaseTime);
-    float getReduction(){return m_curReduction;};
+    void setGainLimit(double newLimit){m_Limit = newLimit;};
+    void setReleaseTime(double releaseTime);
+    double getReduction(){return m_curReduction;};
     int getDelaySamples(){return m_delaySamples;};
 
 private:
@@ -26,20 +26,20 @@ private:
         Rel
     };
 
-    float m_fs;
+    double m_fs;
     int m_nrofchannels;
-    float m_Limit;
-    float m_curReduction;
+    double m_Limit;
+    double m_curReduction;
     int m_delaySamples;
-    float m_releaseTime_ms;
-    float m_alphaRelease;
-    std::vector<std::queue <float>> m_delayline;
-    float m_attackTime_ms;
-    float m_Gain;
+    double m_releaseTime_ms;
+    double m_alphaRelease;
+    std::vector<std::queue <double>> m_delayline;
+    double m_attackTime_ms;
+    double m_Gain;
     int m_attackCounter;
     int m_holdCounter;
 
-    float m_attackIncrement;
+    double m_attackIncrement;
     BrickwallLimiter::State m_state;
 
     void buildAndResetDelayLine();
